[svc] Support coroutine call stacks in TaskService
It's definitely better than to launch an independent task
every time we want to do `co_await` on something.

No proper value returns with `co_return` yet, but that
looks doable if we templatize `CoroSubTaskState`. Still a
special templated awaitable object is needed for that, eww.

As C++ coroutines are kinda awkward to work with, I'd expect
a few more iterations on these interfaces until they converge
to some good balance of usage/implementation complexities.